removed unused variables and renamed gas to fluid
